+++ src/sccz80/const.h
@@ -1,15 +1,15 @@
 /* constant.c */
 extern int constant(LVALUE *lval);
-extern int fnumber(long *val);
+extern int fnumber(int32_t *val);
 extern int stash_double_str(char *start,char *end);
-extern long searchdub(unsigned char *num);
-extern int number(long *val);
+extern int32_t searchdub(unsigned char *num);
+extern int number(int32_t *val);
 extern int hex(char c);
 extern void address(SYMBOL *ptr);
-extern int pstr(long *val);
-extern int tstr(long *val);
-extern int storeq(int length, unsigned char *queue,long *val);
-extern int qstr(long *val);
+extern int pstr(int32_t *val);
+extern int tstr(int32_t *val);
+extern int storeq(int length, unsigned char *queue,int32_t *val);
+extern int qstr(int32_t *val);
 extern void stowlit(int value, int size);
 extern unsigned char litchar(void);
 extern void size_of(LVALUE *lval);
